refactor: make launchbutton it's own component
