+++ src/components/LaunchPage/index.tsx
@@ -1,14 +1,11 @@
-import { YARGStates, YARGVersion } from "@app/hooks/useYARGVersion";
+import { YARGVersion } from "@app/hooks/useYARGVersion";
 import styles from "./styles.module.css";
 import { GenericBox, GenericBoxHeader } from "../GenericBox";
-import Button, { ButtonColor } from "../Button";
-import { DateIcon, InformationIcon, InstallingIcon, LinkIcon, UpdateIcon } from "@app/assets/Icons";
-import PayloadProgress from "../PayloadProgress";
-import { calculatePayloadPercentage } from "@app/utils/Download/payload";
-import { useDialogManager } from "@app/dialogs/DialogProvider";
+import { DateIcon, InformationIcon, LinkIcon } from "@app/assets/Icons";
 import TooltipWrapper from "../TooltipWrapper";
 import { intlFormatDistance } from "date-fns";
 import NewsSection from "../NewsSection";
+import { LaunchButton } from "../LaunchButton";
 
 const INITIAL_RELEASE_DATE = new Date("2023-03-09T05:00:00.000Z");
 
@@ -21,37 +18,10 @@ interface Props {
     icon: string,
 }
 
-interface LaunchButtonProps extends React.PropsWithChildren {
-    style?: React.CSSProperties
-}
-
 const LaunchPage: React.FC<Props> = ({ version, releaseTag, playName, description, websiteUrl, icon }: Props) => {
     // If there isn't a version, something went wrong
     if (!version) {
         return <p>Error: No version.</p>;
-    }
-
-    const dialogManager = useDialogManager();
-
-    function LaunchButton(props: LaunchButtonProps) {
-        if (version.state === YARGStates.NEW_UPDATE) {
-            return <Button style={props.style} color={ButtonColor.GREEN} onClick={() => version.download(dialogManager)}>
-                <UpdateIcon /> Update {playName}
-            </Button>;
-        } else if (version.state === YARGStates.DOWNLOADING) {
-            if (!version.payload) {
-                return <></>;
-            }
-
-            return <Button style={props.style} progress={calculatePayloadPercentage(version.payload)} color={ButtonColor.YELLOW}>
-                <InstallingIcon />
-                <PayloadProgress payload={version.payload} />
-            </Button>;
-        } else {
-            return <Button style={props.style} color={ButtonColor.BLUE} onClick={() => version.play(dialogManager)}>
-                Play {playName}
-            </Button>;
-        }
     }
 
     return <>
@@ -70,7 +40,7 @@ const LaunchPage: React.FC<Props> = ({ version, releaseTag, playName, descriptio
         <div className={styles.main}>
             <NewsSection categoryFilter="yarg" />
             <div className={styles.sidebar}>
-                <LaunchButton style={{ width: "100%" }} />
+                <LaunchButton style={{ width: "100%" }} version={version} playName={playName} />
                 <GenericBox>
                     <GenericBoxHeader>
                         <InformationIcon />
